Move location of client_partial. Give it a default. Use that for proposal properties
