+++ app/helpers/application_helper.rb
@@ -42,14 +42,4 @@ module ApplicationHelper
     controller_name == 'home' ||
     current_page?(carts_path)
   end
-
-  def client_partial(client_name, path, args={})
-    to_check = client_name + "/" + path
-    if lookup_context.template_exists?(to_check, [], true)
-      args[:partial] = to_check
-      render(args)
-    else
-      ""
-    end
-  end
 end
