+++ app/mailers/communicart_mailer.rb
@@ -4,6 +4,7 @@ class CommunicartMailer < ActionMailer::Base
   layout 'communicart_base'
   add_template_helper CommunicartMailerHelper
   add_template_helper TimeHelper
+  add_template_helper ClientHelper
 
 
   def cart_notification_email(to_email, approval)
