+++ app/views/default/_proposal_properties.html.erb
